requirements.txt
google
googleapiclient
cachetools==3.0.0
google-api-python-client==1.7.8
google-api-python-client-helpers==1.2.4
google-auth==1.6.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
httplib2==0.12.0
oauth2client==4.1.3
pyasn1==0.4.5
pyasn1-modules==0.2.3
rsa==4.0
six==1.12.0
uritemplate==3.0.0
docx
python-docx
python-dotenv
